Publish docker images during CI builds (#544)
master builds publish using the git describe output  as the immutable tag and
canary for the mutable tag. Examples:

  gomods/proxy:abc1234
  gomods/proxy:canary

tagged builds publish using the tag as the immutable tag, which is
usually going to be a semver value and with latest for the mutable tag.
Examples:

  gomods/proxy:v1.0.0
  gomods/proxy:latest

branch builds publish using the git describe output as the immutable tag and the
branch name for the mutable tag. Examples:

  gomods/proxy:def4567
  gomods/proxy:beta-1