Add documentation about SSH_AUTH_SOCK from PR#1184 (#1185)
* Add documentation about SSH_AUTH_SOCK from PR#1184

PR#1184 added support for propagating SSH_AUTH_SOCK if it points to a
valid unix socket. As requested, add a section explaining how to use
it.

* auth documentation: replace $(pwd) with $PWD

Replace references to `$(pwd)` with `$PWD` in shell snippets and add
more quoting. This should make these snippets reasonably friendly to
currnet working directory path components containing spaces.

Note: I neither endorse nor recommend spaces in directories, but this is
an easy enough change.

This commit is leaving powershell snippets untouched as I am not
familiar with powershell's word-splitting behavior (if there is any).